gmail-service-3/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage